requirements-railway.txt
fastapi==0.95.1
uvicorn==0.22.0
pandas==2.0.1
numpy==1.23.5
scikit-learn==1.2.2
Pillow==9.5.0
python-multipart==0.0.6
structlog==23.1.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
tensorflow-cpu==2.12.0
google-cloud-storage==2.9.0
google-auth==2.22.0 